Move general parser stuff out of read_csv module and into separate module
